Automatically backed up by Learn

+++ db/migrate/002_create_costume_stores.rb
@@ -6,8 +6,8 @@ class CreateCostumeStores < ActiveRecord::Migration[4.2]
       c.integer :costume_inventory
       c.integer :num_of_employees 
       c.boolean :still_in_business
-      c.float :opening_time
-      c.float :closing_time
+      c.integer :opening_time
+      c.integer :closing_time
     end
   end 
 end
